fix: always prompt for nostr schnorr signing
