+++ src/extension/background-script/router.ts
@@ -124,7 +124,7 @@ const routes = {
       enable: nostr.enable,
       getPublicKeyOrPrompt: nostr.getPublicKeyOrPrompt,
       signEventOrPrompt: nostr.signEventOrPrompt,
-      signSchnorrOrPrompt: nostr.signSchnorrOrPrompt,
+      signSchnorr: nostr.signSchnorr,
       encryptOrPrompt: nostr.encryptOrPrompt,
       decryptOrPrompt: nostr.decryptOrPrompt,
       nip44EncryptOrPrompt: nostr.nip44EncryptOrPrompt,
