+++ src/i18n/locales/en/translation.json
@@ -562,7 +562,7 @@
       "presets": {
         "trust_fully": {
           "title": "I fully trust it",
-          "description": "Auto-sign all requests (except payments)"
+          "description": "Auto-sign all requests (except payments and signing)"
         },
         "reasonable": {
           "title": "Let's be reasonable",
@@ -899,6 +899,7 @@
       "allow_sign_event": "Allow {{host}} to sign a <0>{{kind}}<0></0></0> event",
       "block_and_ignore": "Block and ignore {{host}}",
       "block_added": "Added {{host}} to the blocklist, please reload the website.",
+      "always_confirm": "This request always requires confirmation",
       "recipient": "Recipient",
       "kinds": {
         "unknown": {
